requirements.txt
certifi==2021.10.8
click==8.0.3
cycler==0.11.0
Flask==2.0.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.3
Pillow==8.4.0
pyparsing==3.0.3
python-dateutil==2.8.2
six==1.16.0
torch==1.10.0+cpu
torchvision==0.11.1+cpu
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
